+++ xmtp_db/src/mock.rs
@@ -499,7 +499,7 @@ mock! {
         fn clear_messages<'a>(
             &self,
             group_ids: Option<&'a [Vec<u8>]>,
-            retention_days: Option<i64>,
+            retention_days: Option<u32>,
         ) -> Result<usize, crate::ConnectionError>;
     }
